requirements.txt
pyspark==3.5.1
delta-spark==3.1.0
mysql-connector-python==8.3.0
APScheduler==3.10.4
python-dotenv==1.0.1
pandas==2.2.2
